+++ jovo-integrations/jovo-analytics-googleanalytics/src/index.ts
@@ -1,9 +1,23 @@
 export { GoogleAnalytics } from './GoogleAnalytics';
 export { GoogleAnalyticsAlexa } from './GoogleAnalyticsAlexa';
 export { GoogleAnalyticsGoogleAssistant } from './GoogleAnalyticsGoogleAssistant';
+export { GoogleAnalyticsInstance } from './GoogleAnalyticsInstance';
+export { SystemMetricNamesEnum } from './interfaces';
+import { GoogleAnalyticsGoogleAssistant } from './GoogleAnalyticsGoogleAssistant';
+
+import { Visitor } from 'universal-analytics';
+import { GoogleAnalytics } from './GoogleAnalytics';
 
 // Declare necessary properties for Jovo
-import { Config, Event, Transaction, TransactionItem } from './interfaces';
+import {
+  Config,
+  Event,
+  systemDimensionNames,
+  systemMetricNames,
+  Transaction,
+  TransactionItem,
+} from './interfaces';
+import { GoogleAnalyticsInstance } from './GoogleAnalyticsInstance';
 
 interface AppGoogleAnalyticsConfig {
   GoogleAnalytics?: Config;
@@ -16,31 +30,6 @@ declare module 'jovo-core/dist/src/Interfaces' {
 
 declare module 'jovo-core/dist/src/core/Jovo' {
   interface Jovo {
-    $googleAnalytics: {
-      $data: { [key: string]: string | number };
-      $parameters: { [key: string]: string | number };
-      sendEvent: (params: Event) => void;
-      sendTransaction: (params: Transaction) => void;
-      sendItem: (params: TransactionItem) => void;
-      sendUserEvent: Function;
-      setCustomMetric: (index: number, value: string | number) => void;
-      setCustomDimension: (index: number, value: string | number) => void;
-      /**
-       * Set a custom Google Analytics parameter
-       *
-       * @ref https://developers.google.com/analytics/devguides/collection/protocol/v1/parameters
-       */
-      setParameter: (parameter: string, value: string | number) => void;
-      /**
-       * Set Google Optimize experiment parameters
-       *
-       * @ref https://developers.google.com/optimize/devguides/experiments
-       */
-      setOptimizeExperiment: (experimentId: string, variation: string | number) => void;
-    };
-
-    getRoute(): { intent: string; path: string; type: string };
-
-    getMappedIntentName(): string;
+    $googleAnalytics: GoogleAnalyticsInstance;
   }
 }